// src/interfaces/angularProcessToken.ts
export default interface AngularProcessToken {
  contactId: string;
  officerId: string;
  officerFullName: string;
  contactFullName: string;
  contactAddress: string;
  officerUsername: string;
  contactDOB: string;
  contactMobile: string;
  uprn: string;
  patchId: string;
  subjectId: string;
  areaId: string;
  processId: string;
  IsStarting: boolean;
  HouseholdID: string;
  ManagerId: string;
  TMProcessId: string;
  IncidentId: string;
  ProcessStage: string;
  IsManager: boolean;
  ProcessCRMReference: string;
}